devrequirements.txt
boto3==1.13.19
botocore==1.16.19
click==7.1.2
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
easydict==1.9
Flask==1.1.2
future==0.18.2
imageio==2.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.4
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
PyWavelets==1.1.1
PyYAML==5.3.1
s3transfer==0.3.3
scikit-image==0.17.2
scipy==1.4.1
six==1.15.0
tifffile==2020.5.30
torch==1.5.0
torchvision==0.6.0
urllib3==1.25.9
Werkzeug==0.16.0
